secboot: shuffle code around so that tests with `-tags nosecboot` work correctly
